mandriva: don't fail if host urpmi.cfg contains global options
drop them instead. This forces the user to add them back in the installer.conf file
if some options were required to access the repos.

But it's too much PITA to try to import them, and merge them with the new options
if there's a duplicate since a key can have different name (no-xxx vs xxx name)
and multiple values..