install a minimal /dev in chroot instead of bind mounting the host one
Bind mounting /dev from the host into the chroot is unsecure. Most packages
need a couple of devnodes except the bootloader that requires the disk node
too.

makedev package, which is pulled for no good reasons, creates zillion dev
nodes when it's installed. Since /dev was bind mounted, all nodes are also
created in the host's /dev which is quite bad. Fortunately at the next
reboot, /dev is cleaned since it's a tmpfs is mounted at /dev.